simplification/gathering of all drawing tests into one file
should be easier to browse, and way faster to test
