+++ src/modules/infobox/components/styling/DrawingStyleSizeSelector.vue
@@ -49,7 +49,7 @@ export default {
         },
         dropdownItems() {
             return this.sizes.map((size) => {
-                return new DropdownItem(this.i18n.t(size.label), size)
+                return new DropdownItem(size.label, this.i18n.t(size.label), size)
             })
         },
     },
